fix: turn DatabaseManager of GameService into singleton
This will always use the same dataSource for new requests to actually make use
of the underlying PoolableConnectionFactory in BasicDataSource.
As a result, the service will only maintain one opened connection to the database instead
of opening a new one for each request. This will fix the 'too many clients' error from
the Postgres database which we had in the past.